fix(engine): the fractional rate died at three boundaries downstream of the fix that introduced it
ROUND 2 OF THE AUDIT FOUND A LIVE REGRESSION I SHIPPED THIS MORNING, plus two renderers that undid
the same fix. All three are the same mistake: a value's type was changed at the source and not
followed to the places that consume it.

1. THE SUBPROCESS BOUNDARY - the serious one.

GenStats::rps was changed to publish a fraction below 1/s, because truncating "one request in four
seconds" to 0 says the gateway carried NOTHING. The load generator runs as a CHILD PROCESS and
prints its result as a text stats line; the parent parsed rps with require_i64. "rps=0.25" does not
parse as an i64, so parse_ugen_line classified the ENTIRE WINDOW as Absent::HarnessError.

So the change discarded exactly the rungs it was written to rescue - strictly worse than the false
0 it replaced, because a false 0 is at least a datum a reader can question. Measured:

    rps()=0.25   prints "0.25"   i64 PARSE FAILS -> window becomes HarnessError
    rps()=0.5    prints "0.5"    i64 PARSE FAILS -> window becomes HarnessError
    rps()=19     prints "19"     parses fine
    rps()=44363  prints "44363"  parses fine

Nothing caught it: the engine's unit tests all sit on ONE side of the boundary or the other, and
the format string and the parser live in different files. UgenStats::rps is now f64 via a new
require_f64 - kept BESIDE require_i64 rather than replacing it, because every other field on that
line is a count and a fractional count is a corrupt line that must still be refused. require_f64
also rejects a non-finite rate, so an infinity cannot reach a cast that would SATURATE it into a
plausible-looking 9,223,372,036,854,775,807.

Three tests pin the seam: the round trip for 0.25 / 0.5 / 19 / 44363, a fractional COUNT still
refused, and a non-finite rate refused. Proven red by restoring the i64 parse.

2 and 3. BOTH RENDERERS TRUNCATED IT BACK TO "0".

charts.py's _rate_cell did int(rd["rps"]) and app.js's frontierCell formatted through fmtInt
(Math.round). So even where the artifact carried 0.25 correctly, the page printed "0" -
indistinguishable from a cell that genuinely served nothing, which is the precise statement the
engine had just stopped making. Both now show two decimals below 1/s and whole numbers at or above
it, the same split the engine uses, so the three cannot disagree about what a rate looks like.

ALSO IN THIS COMMIT, from the same round:

MOCK_TTFT_MS was parsed without trimming, three lines above a comment explaining that exact bug for
its neighbours - a value with whitespace made parse() fail, .ok() swallow it, and the mock silently
keep 0 while the operator believed the TTFT they set was in effect.

The THIRD stale `is_rig_bound` reference. mock/src/main.rs still described a 10%-of-ceiling
suppression mechanism that was deleted; run.rs and reverify.rs carried the other two, fixed in the
previous commit. One fact repeated in three comments instead of stated once, and the code beneath
all three changed.

Verified: full CI gate green - fmt, clippy -D warnings, 495 engine tests, all seven python test
files, bench-audit, check-consistency, test-coverage, 201 site tests, charts.py, all four external
auditors, bash -n.

+++ engine/src/run.rs
@@ -719,8 +719,11 @@ pub fn load_window_at(
         parsed.into_value().map(|u| GenStats {
             ok: u.ok.max(0) as u64,
             fail: u.fail.max(0) as u64,
-            elapsed_s: if u.rps > 0 {
-                u.ok as f64 / u.rps as f64
+            // `u.rps` is f64 now (fractional below 1/s), so the comparison and the division are
+            // both float. A sub-1/s window previously could not even reach here: it failed the i64
+            // parse and the whole window was classified as a HarnessError.
+            elapsed_s: if u.rps > 0.0 {
+                u.ok as f64 / u.rps
             } else {
                 0.0
             },
